WasteAiSystem/stablediffusion/requirements.txt
fastapi
uvicorn
openai==0.28.1
torch>=2.0.0
diffusers>=0.25.0
transformers>=4.36.2
accelerate>=0.24.1
matplotlib>=3.7.0
safetensors>=0.4.0
Pillow>=9.0.0
python-consul
python-dotenv